requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
black==24.4.2
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
cryptography==42.0.7
distlib==0.3.8
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identify==2.5.36
idna==3.7
isort==5.13.2
Jinja2==3.1.4
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.9.0
numpy==1.26.4
orjson==3.10.3
packaging==24.0
pandas==2.2.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.2
pre-commit==3.7.1
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.0
uvloop==0.19.0
virtualenv==20.26.2
watchfiles==0.22.0
websockets==12.0
